fix: refactored meeting rater components
